rename: 파이어베이스 목업데이터 data.ts 파일 mock 폴더로 이동
